Add a progress handlebars helper
# Conflicts:
#	templates/actor/character/parts/actor-section-features.hbs
#	templates/actor/partials/actor-clocks.hbs
