+++ module/projectfu.mjs
@@ -80,6 +80,7 @@ import { StashDataModel } from './documents/actors/stash/stash-data-model.mjs';
 import { FUStashSheet } from './sheets/actor-stash-sheet.mjs';
 import { InventoryPipeline } from './pipelines/inventory-pipeline.mjs';
 import { registerKeyBindings } from './keybindings.mjs';
+import { FUHandlebars } from './helpers/handlebars.mjs';
 
 globalThis.projectfu = {
 	ClassFeatureDataModel,
@@ -305,190 +306,7 @@ Hooks.once('setup', () => {});
 /*  Handlebars Helpers                          */
 /* -------------------------------------------- */
 
-// If you need to add Handlebars helpers, here are a few useful examples:
-Handlebars.registerHelper('concat', function () {
-	var outStr = '';
-	for (var arg in arguments) {
-		if (typeof arguments[arg] != 'object') {
-			outStr += arguments[arg];
-		}
-	}
-	return outStr;
-});
-
-Handlebars.registerHelper('toLowerCase', function (str) {
-	return str.toLowerCase();
-});
-
-Handlebars.registerHelper('translate', function (str) {
-	const result = Object.assign(
-		{
-			spell: 'FU.Spell',
-			hp: 'FU.HealthAbbr',
-			mp: 'FU.MindAbbr',
-			ip: 'FU.InventoryAbbr',
-			shields: 'FU.Shield',
-			arcanism: 'FU.Arcanism',
-			chimerism: 'FU.Chimerism',
-			elementalism: 'FU.Elementalism',
-			entropism: 'FU.Entropism',
-			ritualism: 'FU.Ritualism',
-			spiritism: 'FU.Spiritism',
-		},
-		CONFIG.FU.damageTypes,
-		CONFIG.FU.itemTypes,
-		CONFIG.FU.weaponTypes,
-	);
-
-	return result?.[str] ?? str;
-});
-
-Handlebars.registerHelper('getGameSetting', function (settingKey) {
-	return game.settings.get('projectfu', settingKey);
-});
-
-Handlebars.registerHelper('capitalize', function (str) {
-	if (str && typeof str === 'string') {
-		return str.charAt(0).toUpperCase() + str.slice(1);
-	}
-	return str;
-});
-
-Handlebars.registerHelper('uppercase', function (str) {
-	if (str && typeof str === 'string') {
-		return str.toUpperCase();
-	}
-	return str;
-});
-
-Handlebars.registerHelper('neq', function (a, b, options) {
-	if (a !== b) {
-		return options.fn(this);
-	}
-	return '';
-});
-
-Handlebars.registerHelper('half', function (value) {
-	var num = Number(value);
-	if (isNaN(num)) {
-		return '';
-	}
-	return Math.floor(num / 2);
-});
-
-Handlebars.registerHelper('calculatePercentage', function (value, max) {
-	value = parseFloat(value);
-	max = parseFloat(max);
-	const percentage = (value / max) * 100;
-	return percentage.toFixed(2) + '%';
-});
-
-Handlebars.registerHelper('crisis', function (value, max) {
-	value = parseFloat(value);
-	max = parseFloat(max);
-	const half = max / 2;
-	return value <= half;
-});
-
-Handlebars.registerHelper('lookupItemById', function (items, itemId) {
-	return items.find((item) => item._id === itemId);
-});
-
-Handlebars.registerHelper('isItemEquipped', function (item, equippedItems) {
-	if (!item || !equippedItems) {
-		console.error('Item or equippedItems is missing.');
-		return false;
-	}
-	return equippedItems.isEquipped(item);
-});
-
-// Define a Handlebars helper to get the icon class based on item properties
-Handlebars.registerHelper('getIconClass', function (item, equippedItems) {
-	if (!equippedItems) {
-		return '';
-	}
-	return equippedItems.getClass(item);
-});
-
-Handlebars.registerHelper('getSlot', function (item) {
-	if (!item || !item.system) return '';
-	if (item.type === 'weapon') {
-		return item.system.hands.value === 'two-handed' ? 'mainHand' : 'offHand';
-	} else if (item.type === 'shield') {
-		return 'offHand';
-	} else if (item.type === 'armor') {
-		return 'armor';
-	} else if (item.type === 'accessory') {
-		return 'accessory';
-	}
-	return '';
-});
-
-Handlebars.registerHelper('mathAbs', function (value) {
-	return Math.abs(value);
-});
-
-Handlebars.registerHelper('formatMod', function (value) {
-	if (value > 0) {
-		return '+' + value;
-	} else if (value < 0) {
-		return value;
-	}
-	return value;
-});
-
-Handlebars.registerHelper('inArray', function (item, array, options) {
-	if (Array.isArray(array) && array.includes(item)) {
-		return options.fn(this);
-	} else {
-		return options.inverse ? options.inverse(this) : '';
-	}
-});
-
-Handlebars.registerHelper('inSet', function (item, set) {
-	return set.has(item);
-});
-
-Handlebars.registerHelper('formatResource', function (resourceValue, resourceMax, resourceName) {
-	// Convert value to a string to split into 3 digits
-	const valueString = resourceValue.toString().padStart(3, '0');
-	const isCrisis = resourceValue <= resourceMax / 2 && resourceName == 'HP';
-	const digitBoxes = valueString
-		.split('')
-		.map(
-			(digit) =>
-				`<div class="digit-box${isCrisis ? ' crisis' : ''}">
-            <span class="inner-shadow">
-                <span class="number">${digit}</span>
-            </span>
-        </div>`,
-		)
-		.join('');
-
-	return new Handlebars.SafeString(`<span>${resourceName}</span><span class="digit-row">${digitBoxes}</span>`);
-});
-
-Handlebars.registerHelper('math', function (left, operator, right) {
-	left = parseFloat(left);
-	right = parseFloat(right);
-	return {
-		'+': left + right,
-		'-': left - right,
-		'*': left * right,
-		'/': left / right,
-		'%': left % right,
-	}[operator];
-});
-
-Handlebars.registerHelper('includes', function (array, value) {
-	return Array.isArray(array) && array.includes(value);
-});
-
-Handlebars.registerHelper('get', (map, key) => map?.[key]);
-
-Handlebars.registerHelper('clamp', (val, min, max) => {
-	return Math.max(Math.min(val, max), min);
-});
+FUHandlebars.registerHelpers();
 
 /* -------------------------------------------- */
 /*  Ready Hook                                  */
